% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{pegar_nome_formatado}
\alias{pegar_nome_formatado}
\title{Devolve o nome formatado de um indicador}
\usage{
pegar_nome_formatado(indicador, tab_depara)
}
\description{
Devolve o nome formatado de um indicador
}
