% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{pegar_unidade_de_medida}
\alias{pegar_unidade_de_medida}
\title{Devolve a unidade de medida de um indicador}
\usage{
pegar_unidade_de_medida(indicador)
}
\description{
Devolve a unidade de medida de um indicador
}
