% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(auth0 = FALSE, ...)
}
\arguments{
\item{auth0}{whether to use auth0 or not}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
