% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{seleciona_indicadores}
\alias{seleciona_indicadores}
\title{Vetor nomeado com todos os indicadores}
\usage{
seleciona_indicadores(tab_indicadores, tab_depara)
}
\description{
Vetor nomeado com todos os indicadores
}
